requirements.txt
# Core LLM compression
llmlingua>=0.2.2

# Evaluation metrics
bert-score>=0.3.13

# Transformers stack
transformers>=4.40.0
torch>=2.1.0
accelerate>=0.27.0
sentencepiece>=0.1.99

# LLM API clients
openai>=1.30.0
anthropic>=0.25.0

# Arabic NLP utilities
pyarabic>=0.6.15

# Data handling
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and progress
tqdm>=4.65.0

# Development
pytest>=7.4.0
ipykernel>=6.25.0
jupyter>=1.0.0
